add article status management package and workflows for automated updates
