+++ .github/workflows/run_sequential_pipeline.yml
@@ -1,0 +1,37 @@
+name: Run Sequential Pipeline Workflow
+
+on:
+  schedule:
+    # Run every 45 minutes
+    - cron: '*/45 * * * *'
+  workflow_dispatch:  # Allow manual triggering
+
+jobs:
+  run-sequential-pipeline:
+    runs-on: ubuntu-latest
+    
+    steps:
+    - name: Checkout repository
+      uses: actions/checkout@v3
+      
+    - name: Set up Python
+      uses: actions/setup-python@v4
+      with:
+        python-version: '3.10'
+        
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install -r requirements.txt
+        
+    - name: Install Playwright browsers
+      run: |
+        python -m playwright install --with-deps
+        
+    - name: Run Sequential Pipeline script
+      run: python -m createArticles.runSequentialPipeline
+      env:
+        OPENAI_API_KEY: ${{ secrets.OPENAI_API_KEY }}
+        SUPABASE_URL: ${{ secrets.SUPABASE_URL }}
+        SUPABASE_KEY: ${{ secrets.SUPABASE_KEY }}
+        # Add any other environment variables the script needs
