+++ .github/workflows/status_management.yml
@@ -1,0 +1,46 @@
+name: Article Status Management
+
+on:
+  schedule:
+    # Run every 60 minutes
+    - cron: '0 * * * *'
+  workflow_dispatch:  # Allow manual triggering of the workflow
+
+jobs:
+  update-article-statuses:
+    runs-on: ubuntu-latest
+    env:
+      SUPABASE_URL: ${{ secrets.SUPABASE_URL }}
+      SUPABASE_KEY: ${{ secrets.SUPABASE_KEY }}
+      
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v4
+      
+      - name: Set up Python
+        uses: actions/setup-python@v4
+        with:
+          python-version: '3.12'
+          cache: 'pip'
+      
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install -r requirements.txt
+      
+      - name: Run status management functions
+        run: |
+          python -c "
+          from createArticles.statusManagement import update_article_statuses, update_missing_statuses, cleanup_archived_articles
+          
+          print('Running update_missing_statuses...')
+          update_missing_statuses()
+          
+          print('Running cleanup_archived_articles...')
+          cleanup_archived_articles()
+          
+          print('Running update_article_statuses...')
+          update_article_statuses()
+          
+          print('Status management completed successfully')
+          "
