+++ createArticles/storeInDB.py
@@ -3,6 +3,8 @@ from urllib.parse import urlparse
 from datetime import datetime, timedelta, timezone
 from createArticles.detectTeam import detectTeam
 from createArticles.fetchUnprocessedArticles import get_unprocessed_articles, supabase_client
+# Import functions from their new location
+from createArticles.statusManagement import update_article_statuses, update_missing_statuses, cleanup_archived_articles
 
 def create_news_article_record(
     article: dict,
@@ -106,158 +108,6 @@ def process_articles(unprocessed_articles, english_articles, german_articles, im
         # If insertion succeeded, mark the article as processed
         if new_record_id:
             mark_article_as_processed(article_id)
-
-def update_missing_statuses():
-    """
-    Updates articles that don't have a status set (NULL).
-    Sets their status based on their age:
-    - Articles older than 2 hours -> "OLD"
-    - Articles older than 48 hours -> "ARCHIVED"
-    - Articles newer than 2 hours -> "NEW"
-    """
-    try:
-        # Get current time in UTC
-        now = datetime.now(timezone.utc)
-        two_hours_ago = now - timedelta(hours=2)
-        two_days_ago = now - timedelta(days=2)
-
-        # Fetch all articles with NULL status
-        response = supabase_client.table("NewsArticle")\
-            .select("id", "created_at")\
-            .is_("Status", "null")\
-            .execute()
-
-        if not response.data:
-            print("No articles found with missing status.")
-            return
-
-        for article in response.data:
-            article_id = article['id']
-            # Parse the timestamp and make it timezone-aware
-            created_at = datetime.fromisoformat(article['created_at'].replace('Z', '+00:00'))
-            
-            # Determine status based on age
-            if created_at < two_days_ago:
-                new_status = "ARCHIVED"
-            elif created_at < two_hours_ago:
-                new_status = "OLD"
-            else:
-                new_status = "NEW"
-
-            # Update article status
-            supabase_client.table("NewsArticle")\
-                .update({"Status": new_status})\
-                .eq("id", article_id)\
-                .execute()
-            
-            print(f"Set status of article {article_id} to {new_status}")
-
-    except Exception as e:
-        print(f"Error updating missing statuses: {e}")
-
-def update_article_statuses():
-    """
-    Updates the status of articles based on their age:
-    - Articles older than 2 hours -> "OLD"
-    - Articles older than 48 hours -> "ARCHIVED"
-    """
-    try:
-        # Get current time in UTC
-        now = datetime.now(timezone.utc)
-        two_hours_ago = now - timedelta(hours=2)
-        two_days_ago = now - timedelta(days=2)
-
-        # First update any articles with missing status
-        update_missing_statuses()
-
-        # Then update existing statuses
-        # Fetch all articles that are not ARCHIVED and have a status
-        response = supabase_client.table("NewsArticle")\
-            .select("id", "created_at", "Status")\
-            .not_.eq("Status", "ARCHIVED")\
-            .not_.is_("Status", "null")\
-            .execute()
-
-        if not response.data:
-            print("No articles found to update.")
-            return
-
-        for article in response.data:
-            article_id = article['id']
-            # Parse the timestamp and make it timezone-aware
-            created_at = datetime.fromisoformat(article['created_at'].replace('Z', '+00:00'))
-            current_status = article['Status']
-            
-            # Don't change UPDATED status articles unless they're old enough to be archived
-            if current_status == "UPDATED" and created_at >= two_days_ago:
-                continue
-                
-            # Determine new status based on age
-            if created_at < two_days_ago:
-                new_status = "ARCHIVED"
-            elif created_at < two_hours_ago:
-                new_status = "OLD"
-            else:
-                continue  # Skip if article is newer than 2 hours
-
-            # Update article status
-            supabase_client.table("NewsArticle")\
-                .update({"Status": new_status})\
-                .eq("id", article_id)\
-                .execute()
-            
-            print(f"Updated article {article_id} status to {new_status}")
-
-    except Exception as e:
-        print(f"Error updating article statuses: {e}")
-
-def cleanup_archived_articles():
-    """
-    One-time cleanup function to fix articles that were incorrectly marked as ARCHIVED
-    when using sourceArticlePublishedAt instead of created_at.
-    """
-    try:
-        # Get current time in UTC
-        now = datetime.now(timezone.utc)
-        two_hours_ago = now - timedelta(hours=2)
-        two_days_ago = now - timedelta(days=2)
-
-        # Fetch all articles
-        response = supabase_client.table("NewsArticle")\
-            .select("id", "created_at", "Status")\
-            .execute()
-
-        if not response.data:
-            print("No articles found.")
-            return
-
-        fixed_count = 0
-        for article in response.data:
-            article_id = article['id']
-            created_at = datetime.fromisoformat(article['created_at'].replace('Z', '+00:00'))
-            current_status = article.get('Status')
-            
-            # Determine correct status based on created_at
-            if created_at < two_days_ago:
-                new_status = "ARCHIVED"
-            elif created_at < two_hours_ago:
-                new_status = "OLD"
-            else:
-                new_status = "NEW"
-
-            # Only update if status is different
-            if current_status != new_status:
-                supabase_client.table("NewsArticle")\
-                    .update({"Status": new_status})\
-                    .eq("id", article_id)\
-                    .execute()
-                print(f"Fixed article {article_id}: {current_status} -> {new_status}")
-                fixed_count += 1
-
-        print(f"\nCleanup complete. Fixed {fixed_count} articles.")
-
-    except Exception as e:
-        print(f"Error during cleanup: {e}")
 
 if __name__ == '__main__':
     # Load all necessary data
